+++ CHANGELOG.md
@@ -24,7 +24,7 @@ FlatLaf Change Log
   #718)
 - TextField: Fixed placeholder text painting, which did not respect horizontal
   alignment property of `JTextField`. (issue #721)
-- Popop: Fixed drop shadow if popup overlaps a heavyweight component. (Windows
+- Popup: Fixed drop shadow if popup overlaps a heavyweight component. (Windows
   10 only; issue #626)